+++ tests/testnet/ackerman.ts
@@ -1,25 +1,24 @@
-import { inputIndex, inputSatoshis } from './util/txHelper'
+import {
+    inputIndex,
+    inputSatoshis,
+    testnetDefaultSigner,
+} from './util/txHelper'
 import { Ackermann } from '../../src/contracts/ackermann'
-import { bsv, TestWallet, WhatsonchainProvider } from 'scrypt-ts'
-import { myPrivateKey } from './util/myPrivateKey'
+import { bsv } from 'scrypt-ts'
 
 async function main() {
     await Ackermann.compile()
     const ackermann = new Ackermann(2n, 1n)
 
-    const signer = new TestWallet(myPrivateKey).connect(
-        new WhatsonchainProvider(bsv.Networks.testnet)
-    )
-
     // connect to a signer
-    ackermann.connect(signer)
+    ackermann.connect(testnetDefaultSigner)
 
     // contract deploy
     const deployTx = await ackermann.deploy(inputSatoshis)
     console.log('Ackermann contract deployed: ', deployTx.id)
 
     // contract call
-    const changeAddress = await signer.getDefaultAddress()
+    const changeAddress = await testnetDefaultSigner.getDefaultAddress()
     const unsignedCallTx: bsv.Transaction = await new bsv.Transaction()
         .addInputFromPrevTx(deployTx)
         .change(changeAddress)
@@ -32,7 +31,9 @@ async function main() {
                 cloned.unlock(5n)
             })
         })
-    const callTx = await signer.signAndsendTransaction(unsignedCallTx)
+    const callTx = await testnetDefaultSigner.signAndsendTransaction(
+        unsignedCallTx
+    )
     console.log('Ackermann contract called: ', callTx.id)
 }
