+++ tests/testnet/util/privateKey.ts
